requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector Search and Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
huggingface-hub==0.19.4

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
python-magic==0.4.27
chardet==5.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# LLM Support
transformers==4.36.0
torch==2.1.2
openai==1.3.7
anthropic==0.7.7
tokenizers==0.15.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Optional: Streamlit for UI
streamlit==1.29.0
plotly==5.18.0
